requirements.txt
pillow
py7zr